feat: skip build and reuse previous build artifacts
